+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+
+- Add `highlight()` and `annotateText()` for literal text annotations.
+
 ## [0.7.0] - 2026-09-20
 
 - Initial release.